local-services/ui-service/requirements.txt
fastapi[standard]
google-cloud-bigquery